feat: unifique interaction/message (not full tested)
